Fix CodeQL rate-limiting alert and restore /api/reframe route.
- Re-implemented `express-rate-limit` for `/api/reframe` and other routes.
- Restored the `/api/reframe` route which was accidentally removed.
- Maintained security updates for `axios` and `react-router-dom`.
