+++ package.json
@@ -22,6 +22,7 @@
     "cors": "^2.8.5",
     "dotenv": "^17.2.3",
     "express": "^5.2.1",
+    "express-rate-limit": "^7.5.0",
     "gemini-mcp-tool": "^1.1.4",
     "gemini-zod": "^0.1.2",
     "jsonwebtoken": "^9.0.3",
