--- server.log
@@ -1,1 +1,0 @@
-Server is running on port 8080.
