+++ server/index.js
@@ -8,11 +8,13 @@ const { getReframeSuggestion } = require("./reframe");
 
 const app = express();
 
-const reframeLimiter = rateLimit({
+const limiter = rateLimit({
   windowMs: 15 * 60 * 1000, // 15 minutes
   max: 100, // Limit each IP to 100 requests per windowMs
-  message: "Too many requests from this IP, please try again later."
 });
+
+// Apply rate limiting to all requests
+app.use(limiter);
 
 // CORS and Body Parsing
 app.use(cors());
@@ -46,7 +48,7 @@ app.get("/api/test/admin", [verifyToken, isAdmin], (req, res) => {
   res.status(200).send("Admin Content.");
 });
 
-app.post("/api/reframe", [verifyToken, reframeLimiter], async (req, res) => {
+app.post("/api/reframe", [verifyToken], async (req, res) => {
   const { distortion, text } = req.body;
   if (!distortion || !text) {
     return res.status(400).send({ message: "Distortion and text are required." });
