+++ src/nlp/cd-detector.js
@@ -54,7 +54,7 @@ async function generateReferenceEmbeddings() {
 
 // Function to calculate cosine similarity between two tensors
 function cosineSimilarity(vec1, vec2) {
-  return tf.losses.cosineDistance(vec1, vec2, -1).neg().add(1);
+  return tf.metrics.cosineDistance(vec1, vec2).neg().add(1);
 }
 
 // sensitivity: 'low', 'medium', 'high'
@@ -93,3 +93,8 @@ async function detectCDs(text, sensitivity = 'medium') {
 }
 
 export { detectCDs };
+
+// Initial model load
+if (process.env.NODE_ENV !== 'test') {
+  loadModel();
+}
