+++ src/nlp/cd-detector.test.js
@@ -21,8 +21,8 @@ jest.mock('@tensorflow/tfjs', () => {
   return {
     ...originalTf,
     mean: jest.fn((tensor) => mockTensor), // Always return mockTensor
-    losses: {
-      ...originalTf.losses,
+    metrics: {
+      ...originalTf.metrics,
       cosineDistance: jest.fn((vec1, vec2) => createMockSqueezedTensor(0.9)), // Default high similarity
     },
     tensor2d: jest.fn(() => mockTensor),
@@ -55,7 +55,7 @@ describe('detectCDs', () => {
 
   test('should return an empty array if no cognitive distortions are detected', async () => {
     // Simulate low similarity for all CDs
-    tf.losses.cosineDistance.mockReturnValue(createMockSqueezedTensor(0.1));
+    tf.metrics.cosineDistance.mockReturnValue(createMockSqueezedTensor(0.1));
     const text = 'This is a neutral sentence.';
     const result = await detectCDs(text);
     expect(result).toEqual([]);
@@ -68,7 +68,7 @@ describe('detectCDs', () => {
     // This requires knowing the order of CD types being checked in detectCDs
     // For simplicity, we'll make all cosine distances high enough to trigger detection
     // and then filter based on the expected outcome.
-    tf.losses.cosineDistance.mockImplementation((vec1, vec2) => {
+    tf.metrics.cosineDistance.mockImplementation((vec1, vec2) => {
       // In a real scenario, you might inspect vec1/vec2 to determine which CD is being checked
       // For now, we'll just return a high similarity for all.
       return createMockSqueezedTensor(0.95);
